Create Core code subroutine
The competition track has a white surface, with continuous black lines on both sides serving as guiding lines, where the black line width is 25mm. The vehicle model detects the path by collecting images of the track, following the competition rules that specify the width of the track zone to be 40cm, the minimum corner radius to be 50cm, and the specific indicators of various track markers. The goal of the participating teams is to enable the smart car to complete a lap around the track in the shortest time according to the rules.According to the relevant competition rules, this smart car system uses the C-type vehicle model provided uniformly by the competition organizing committee and employs a 32-bit microcontroller TC377 produced by Infineon as its core controller, with software development conducted in the ADS development environment. The track images are captured by a CMOS image sensor, and the image signals are input to the TC377 microcontroller for further processing to obtain key track information; speed is detected using encoders, and orthogonal decoding is performed by a microcontroller for pulse calculations to acquire speed and distance. Differential PD control is implemented; a PI controller drives the motor, adjusting its power through PWM control; the target speed is comprehensively controlled by default settings, operational safety plans, and optimization strategies based on image processing. Given the various elements in this competition, adding sensors is necessary for better data collection on track features; we chose an infrared module for distance measurement and utilized a gyroscope to obtain the vehicle's posture for entering and exiting the garage. Additionally, two horizontal inductors were used to enhance the stability of roundabout judgment.